ECA-8672: Remove unnecessary boxing/unboxing. Fix some variable names
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@33806 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/cesecore-common/src-test/org/cesecore/certificates/endentity/ExtendedInformationTest.java
@@ -151,8 +151,8 @@ public class ExtendedInformationTest {
         final String maxFailedLoginAttempts = "2";
         extendedInformation.getRawData().put("remainingloginattempts", remainingLoginAttempts);
         extendedInformation.getRawData().put("maxfailedloginattempts", maxFailedLoginAttempts);
-        assertEquals(Integer.valueOf(remainingLoginAttempts), (Integer) extendedInformation.getRemainingLoginAttempts());
-        assertEquals(Integer.valueOf(maxFailedLoginAttempts), (Integer) extendedInformation.getMaxLoginAttempts());
+        assertEquals(remainingLoginAttempts, extendedInformation.getRemainingLoginAttempts());
+        assertEquals(maxFailedLoginAttempts, extendedInformation.getMaxLoginAttempts());
     }
 
     @Test
